requirements.txt
beautifulsoup4
pandas
requests
python-whois